Versão 4.17.35: Melhora o desempenho no processo de desconstruir polígonos
