+++ DsgTools/core/DSGToolsProcessingAlgs/dsgtoolsProcessingAlgorithmProvider.py
@@ -129,9 +129,6 @@ from DsgTools.core.DSGToolsProcessingAlgs.Algs.ValidationAlgs.snapFeaturesInside
 from DsgTools.core.DSGToolsProcessingAlgs.Algs.ValidationAlgs.splitContoursAtMaximumLengthAlgorithm import (
     SplitContoursAtMaximumLengthAlgorithm,
 )
-from DsgTools.core.DSGToolsProcessingAlgs.Algs.ValidationAlgs.unbuildPolygonsAlgorithmV2 import (
-    UnbuildPolygonsAlgorithmV2,
-)
 from processing.core.ProcessingConfig import ProcessingConfig, Setting
 from PyQt5.QtCore import QCoreApplication
 from qgis.core import QgsApplication, QgsProcessingProvider
@@ -789,7 +786,6 @@ class DSGToolsProcessingAlgorithmProvider(QgsProcessingProvider):
             ReclassifyGroupsOfPixelsToNearestNeighborAlgorithmV2(),
             ReclassifyGroupsOfPixelsToNearestNeighborAlgorithmV3(),
             LineOnAreaOverlayerAlgorithm(),
-            UnbuildPolygonsAlgorithmV2(),
             PolygonTilerAlgorithm(),
         ]
         return algList
