+++ DsgTools/metadata.txt
@@ -10,7 +10,7 @@
 name=DSG Tools
 qgisMinimumVersion=3.22
 description=Brazilian Army Cartographic Production Tools
-version=4.17.34
+version=4.17.35
 author=Brazilian Army Geographic Service
 email=dsgtools@dsg.eb.mil.br
 about=
